+++ rer/groupware/room/config.py
@@ -3,10 +3,5 @@
 PROJECTNAME = 'rer.groupware.room'
 
 ADD_PERMISSIONS = {
-    'GroupRoom': 'rer.groupware.room: Add GroupRoom',
-#    'DocumentsArea': 'rer.groupware.room: Add DocumentsArea',
-#    'EventsArea': 'rer.groupware.room: Add EventsArea',
-#    'NewsArea': 'rer.groupware.room: Add NewsArea',
-#    'PollsArea': 'rer.groupware.room: Add PollsArea',
-#    'ProjectsArea': 'rer.groupware.room: Add ProjectsArea',
+    'GroupRoom': 'rer.groupware.room: Add GroupRoom'
 }